requirements.txt
zc.lockfile==2.0
bmp280==0.0.4
mh_z19==3.1.3
python-dotenv==0.20.0
paho-mqtt==1.6.1
PyP100==0.0.19
